THE FINAL PRODUCT
So there's a problem with the actual ordering portion of the program, and we are aware of it. However, we cannot figure out a fix to it so in the main cpp file, we've labeled where the program bricks, what causes it to brick and what should happen in magical christmasland.
Also apparently magical christmasland is a term only really used by mtg players. It means a scenario where everything works perfectly, but will logically never happen.